Moving advancement methods out of utils
