+++ module/documents/actor.mjs
@@ -1,7 +1,8 @@
 import { Dice } from "../dice.mjs";
 import { RollDialog } from "../helpers/roll-dialog.mjs";
 import { resizeTokens } from "../helpers/actor.mjs";
-import { getItemsOfType, roleValueChange } from "../helpers/utils.mjs";
+import { roleValueChange } from "../helpers/advancement.mjs";
+import { getItemsOfType } from "../helpers/utils.mjs";
 
 /**
  * Extend the base Actor document by defining a custom roll data structure which is ideal for the Simple system.
