+++ module/helpers/utils.mjs
@@ -275,152 +275,3 @@ export function deleteAttachmentsForItem(item, actor, previousLevel=null) {
     }
   }
 }
-
-/**
- * @param {Object} actor The acrot that the role is attached to
- * @param {Array} arrayLevels An array of the levels at which a value changes
- * @param {Number} lastProcessedLevel The value of actor.system.level the last time a level change was processed
- * @returns {Number} totalChange The number of level changes.
- */
-export function roleValueChange(currentLevel, arrayLevels, lastProcessedLevel=null) {
-  const levelDiff = currentLevel - lastProcessedLevel;
-  if (!levelDiff) {
-    return 0;
-  }
-
-  const isLevelUp = currentLevel > lastProcessedLevel;
-  const changeIncrement = isLevelUp ? 1 : -1;
-  let totalChange = 0;
-
-  for (const arrayLevel of arrayLevels) {
-    const level = arrayLevel.replace(/[^0-9]/g, '');
-    const actorReachedLevel =
-          isLevelUp && level <= currentLevel
-      || !isLevelUp && level > currentLevel;
-    const levelNotYetProcessed =
-      !lastProcessedLevel
-      || (isLevelUp  && level >  lastProcessedLevel)
-      || (!isLevelUp && level <= lastProcessedLevel);
-    const valueChange = actorReachedLevel && levelNotYetProcessed;
-    totalChange += valueChange ? changeIncrement : 0;
-  }
-
-  return totalChange;
-}
-
-/**
- * Handles setting the values and items for an actor's role
- * @param {Object} role The actor's Role
- * @param {Actor} actor The actor
- * @param {Number} newLevel (Optional) The new level that you are changing to
- * @param {Number} previousLevel (Optional) The last level processed for the actor
- */
-export async function setRoleValues(role, actor, newLevel=null, previousLevel=null) {
-  for (const essence in role.system.essenceLevels) {
-    const totalChange = roleValueChange(actor.system.level, role.system.essenceLevels[essence], previousLevel);
-    const essenceValue = actor.system.essences[essence] + totalChange;
-    const essenceString = `system.essences.${essence}`;
-
-    await actor.update({
-      [essenceString]: essenceValue,
-    });
-  }
-
-  if (role.system.powers.personal.starting) {
-    const totalChange = roleValueChange(actor.system.level, role.system.powers.personal.levels, previousLevel);
-    let newPersonalPowerMax = 0;
-    if (actor.system.powers.personal.max > 0) {
-      newPersonalPowerMax = actor.system.powers.personal.max + role.system.powers.personal.increase * totalChange;
-    } else {
-      newPersonalPowerMax = role.system.powers.personal.starting;
-    }
-
-    await actor.update({
-      "system.powers.personal.max": newPersonalPowerMax,
-      "system.powers.personal.regeneration": role.system.powers.personal.regeneration,
-    });
-  }
-
-  if (role.system.adjustments.health.length) {
-    const totalChange = roleValueChange(actor.system.level, role.system.adjustments.health, previousLevel);
-    const newHealthBonus = actor.system.health.bonus + totalChange;
-
-    await actor.update({
-      "system.health.bonus": newHealthBonus,
-    });
-  }
-
-  if (role.system.skillDie.isUsed) {
-    const skillName = "roleSkillDie";
-    const shiftList = CONFIG.E20.skillShiftList;
-    const totalChange = roleValueChange(actor.system.level, role.system.skillDie.levels, previousLevel);
-    let initialShiftIndex = shiftList.findIndex(s => s == "d2");
-    if (actor.system.skills[skillName].shift) {
-      initialShiftIndex = shiftList.findIndex(s => s == actor.system.skills[skillName].shift);
-    }
-
-    const finalShiftIndex = Math.max(
-      0,
-      Math.min(shiftList.length - 1, initialShiftIndex-totalChange),
-    );
-
-    const skillStringShift = `system.skills.${skillName}.shift`;
-    const skillStringIsSpecialized = `system.skills.${skillName}.isSpecialized`;
-
-    let isSpecialized = false;
-    for (const arrayLevel of role.system.skillDie.specializedLevels) {
-      const level = arrayLevel.replace(/[^0-9]/g, '');
-      if (actor.system.level == level) {
-        isSpecialized = true;
-        break;
-      }
-    }
-
-    await actor.update({
-      [skillStringShift]: shiftList[finalShiftIndex],
-      [skillStringIsSpecialized] : isSpecialized,
-    });
-  }
-
-  for (const [,item] of Object.entries(role.system.items)) {
-    if (item.type == "rolePoints" && actor.flags.essence20.roleDrop) {
-      await createItemCopies(role.system.items, actor, "rolePoints", role);
-      // Set points value to max?
-    }
-  }
-
-  if (newLevel && previousLevel && newLevel > previousLevel || (!newLevel && !previousLevel)) {
-    // Drop or level up
-    await createItemCopies(role.system.items, actor, "perk", role, previousLevel);
-  } else {
-    // Level down
-    await deleteAttachmentsForItem(role, actor, previousLevel);
-  }
-
-  actor.setFlag('essence20', 'roleDrop', false);
-}
-
-/**
- * Handles setting the values and items for an actor's focus
- * @param {Focus} focus The actor's Focus
- * @param {Actor} actor The actor
- * @param {Number} newLevel (Optional) The new level that you are changing to
- * @param {Number} previousLevel (Optional) The last level processed for the actor
- */
-export async function setFocusValues(focus, actor, newLevel=null, previousLevel=null) {
-  const totalChange = roleValueChange(actor.system.level, focus.system.essenceLevels, previousLevel);
-  const essenceValue = actor.system.essences[actor.system.focusEssence] + totalChange;
-  const essenceString = `system.essences.${actor.system.focusEssence}`;
-
-  await actor.update({
-    [essenceString]: essenceValue,
-  });
-
-  if (newLevel && previousLevel && newLevel > previousLevel || (!newLevel && !previousLevel)) {
-    // Drop or level up
-    await createItemCopies(focus.system.items, actor, "perk", focus, previousLevel);
-  } else {
-    // Level down
-    await deleteAttachmentsForItem(focus, actor, previousLevel);
-  }
-}
